Drop student_events table only if it exists.
We'll rename it before we run the migrations in production,
so it won't exist there.

+++ db/migrate/20140112023830_drop_student_events.rb
@@ -1,6 +1,6 @@
 class DropStudentEvents < ActiveRecord::Migration
   def up
-    drop_table :student_events
+    drop_table :student_events if connection.table_exists? :student_events
   end
 
   def down
